% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{as.matrix.cells}
\alias{as.matrix.cells}
\title{Convert Cells Object to Matrix}
\usage{
\method{as.matrix}{cells}(x, ...)
}
\arguments{
\item{x}{A Cells object.}
}
\value{
A matrix representing the cell values.
}
\description{
Converts a valid Cells object to a matrix. Ensures structure is appropriate
for conversion and arranges data accordingly.
}
\keyword{internal}
