% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_header.R
\name{attach_header}
\alias{attach_header}
\title{Attach header meta-data to data by specified direction}
\usage{
attach_header(
  dat,
  hdr,
  direction = c("v", "h", "vl", "vr", "vm", "hu", "hd", "hm", "direct", "nothing")
)
}
\arguments{
\item{dat}{Data frame with data cells (must have 'row', 'col')}

\item{hdr}{Data frame with header cells (must have 'row' or 'col' as relevant)}

\item{direction}{Header direction, one of:
"v" (vertical), "h" (horizontal),
"vl", "vr", "vm" (vertical: left, right, mid),
"hu", "hd", "hm" (horizontal: up, down, mid),
"direct" (universal)}
}
\value{
Data frame with attached header columns (with "_hdr" suffix)
}
\description{
Attach header meta-data to data by specified direction
}
\keyword{internal}
