% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/value_attribute_classifiers.R
\name{basic_classifier}
\alias{basic_classifier}
\alias{basic_classifier.default}
\alias{basic_classifier.cell_df}
\title{Naive \code{data_type} Based Value/Attribute Classifier}
\usage{
basic_classifier(x, ...)

\method{basic_classifier}{default}(x, ...)

\method{basic_classifier}{cell_df}(x, ...)
}
\arguments{
\item{x}{Either a \code{\link[=cell_df-class]{cell_df}} or leave blank for function output
(required by \code{\link[=value_attribute_classify]{value_attribute_classify}})}
}
\description{
Naive \code{data_type} Based Value/Attribute Classifier
}
\seealso{
\link[=value_attribute_classify]{value_attribute_classify}
}
\keyword{internal}
