% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-cell_df_analysis.R
\name{cell_analysis-class}
\alias{cell_analysis-class}
\title{\code{cell_analysis} class}
\description{
The \code{cell_analysis} class is based on list,
created in order to store analysis of cell level information (\code{\link[=cell_df-class]{cell_df}}).
}
\section{Properties of \code{cell_analysis}}{


Objects of class \code{cell_analysis} have following named nodes:
\itemize{
\item \code{cells} : Contains information about \code{cell_group_type} in terms of (data, minor and major attributes).(a tibble)
\item \code{sections} : Contains boundaries of each data block. (a tibble)
\item \code{details} : a list containing further information
\item \code{cell_df} : The original cell_df which is passed for processing
}
}

\section{Applicable methods on \code{cell_analysis}}{

\itemize{
\item \code{print}: Prints identifier of the class and the number of blocks (and potential issues if any).
\item \code{plot}: Plots (using \code{\link[ggplot2:ggplot]{ggplot2}}) the data-block information.
}
}

