% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cell_trace_plot.R
\name{cell_composition_traceback}
\alias{cell_composition_traceback}
\title{Display static composition traceback plot}
\usage{
cell_composition_traceback(ca, dc, trace_row = 1L)
}
\arguments{
\item{ca}{A cell analysis}

\item{dc}{(optional) a composed tidy data.frame}

\item{trace_row}{the row number to trace. (Default 1)}
}
\description{
Display static composition traceback plot
}
\keyword{internal}
