% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_columns_public.R
\name{collate_columns.cells_analysis}
\alias{collate_columns.cells_analysis}
\title{Collate columns of a cells_analysis object}
\usage{
\method{collate_columns}{cells_analysis}(
  x,
  ...,
  fixed_columns = c("row", "col", "data_gid", "value"),
  similarity_threshold = 0,
  dicard_cell_address = TRUE,
  detect_data_block_info_cols = TRUE
)
}
\arguments{
\item{x}{A cells_analysis object, which is a specialized data structure for
analyzing cell compositions.}

\item{...}{Additional arguments passed to the function.}

\item{fixed_columns}{A character vector of column names that should not be
altered during the collation process. Default is \code{c("row", "col", "data_gid", "value")}.}

\item{similarity_threshold}{A numeric value indicating the minimum similarity
score required for two columns to be considered similar. Default is \code{0}. (0
means it will try to reduce columns as much as possible)}

\item{dicard_cell_address}{A logical value indicating whether to discard cell
address columns (like \code{row}, \code{col}, \code{data_gid}) from the output. Default is
\code{TRUE}.}
}
\value{
A data frame with collated columns, where similar columns are merged
based on their similarity scores.
}
\description{
This function collates columns of a cells_analysis object, which is a
specialized data structure for analyzing cell compositions. It merges similar
columns based on a similarity threshold, allowing for flexible handling of
data.
}
\keyword{internal}
