% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compose_public.R
\name{compose}
\alias{compose}
\title{Compose and Stitch Cells from a Cell-Analysis Object}
\usage{
compose(ca, simplify = TRUE)
}
\arguments{
\item{ca}{A \code{cell_analysis} object containing cell attribute mapping and data
block information.}

\item{simplify}{Logical. If \code{TRUE} (default), returns a single data frame of
stitched cells. If \code{FALSE}, returns a list of data frames (composition of
cells) for further processing (like collate columns).}
}
\value{
If \code{simplify = TRUE}, a data frame of stitched cells; otherwise, a
list of data frames of class \code{core_cells_composition_class}.
}
\description{
This function reconstructs ("composes") by stitching data/value cells with
attribute cells from a given cell-analysis object. It combines attribute
mapping and data blocks to produce a data frame (or a list of data frames)
representing the structured, semantically-stitched table.
}
\details{
This function expects a valid \code{cell_analysis} object produced by the
table structure learning pipeline. It combines attribute-data mappings with
data blocks, joining cell values and header information according to the
mapping and orientation tags.

If the object does not contain attribute-data mapping, an invisible \code{-1} is
returned and a warning is issued.
}
