% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_helpers.R
\name{core_as_cells_internal}
\alias{core_as_cells_internal}
\title{Internal function to convert data frames to cells format based on
intermediate structure.}
\usage{
core_as_cells_internal(df, ...)
}
\description{
Internal function to convert data frames to cells format based on
intermediate structure.
}
\keyword{internal}
