% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{core_as_rc_df}
\alias{core_as_rc_df}
\title{Convert a \code{cells} Object to a \code{rc_df} Format}
\usage{
core_as_rc_df(x)
}
\description{
This is basically tibble with only row and col columns.
}
\keyword{internal}
