% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_helpers.R
\name{core_as_sheets_internal}
\alias{core_as_sheets_internal}
\title{Internal function to convert list or data.frames to sheets format based on
intermediate structure.}
\usage{
core_as_sheets_internal(lst, ...)
}
\description{
Internal function to convert list or data.frames to sheets format based on
intermediate structure.
}
\keyword{internal}
