% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{core_cells_validation_end_use}
\alias{core_cells_validation_end_use}
\title{Validate a \code{cells} Object and Optionally Halt Execution}
\usage{
core_cells_validation_end_use(dat)
}
\arguments{
\item{dat}{An R object to validate against the \code{cells} definition.}
}
\value{
An logical value indicating whether the validation passed
}
\description{
This function checks if the provided \code{cells} object is valid according to the
\code{core_validate_cells()} function. If the validation fails, it prints error
messages and can halt execution if the \code{fail} argument is set to \code{TRUE}.
Ideally this function should be used everywhere apart from this file
}
\keyword{internal}
