% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{core_is_rcdf}
\alias{core_is_rcdf}
\title{Internal Generic for Checking if an Object is a Valid \code{rc_df}}
\usage{
core_is_rcdf(d)
}
\description{
Internal Generic for Checking if an Object is a Valid \code{rc_df}
}
\keyword{internal}
