% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_preprocess.R
\name{core_prep_check_str_categorical}
\alias{core_prep_check_str_categorical}
\title{Check if a character vector contains categorical values based on frequent
words.}
\usage{
core_prep_check_str_categorical(x, allowed_strings = NULL)
}
\arguments{
\item{x}{A character vector to be checked.}

\item{allowed_strings}{A character vector of strings to be used as patterns
for matching. If not provided, the function will use the most frequent
words from \code{x} as patterns.}
}
\value{
A character vector of the same length as \code{x} with values
"categorical", "non_categorical", or "blank".
}
\description{
This function evaluates each element of a character vector \code{x} to determine
if it matches any of the provided frequent words or patterns.
}
\keyword{internal}
