% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_preprocess.R
\name{core_prep_check_str_date}
\alias{core_prep_check_str_date}
\title{Check if a character vector contains date formats.}
\usage{
core_prep_check_str_date(
  x,
  allowed_strings = NULL,
  use_date_regex = c("base", "base_with_qtrs", "base_with_qtrs_with_fy",
    "base_with_qtrs_with_fy_extra")
)
}
\arguments{
\item{x}{A character vector to be checked.}

\item{allowed_strings}{A character vector of strings to be ignored during the
check. (These are called "allowed strings" as they are considered part of
a date if supplied with a date expression, e.g., "DOB:")}

\item{use_date_regex}{A character string indicating the regex pattern to use}
}
\value{
A character vector of the same length as \code{x} with values "date",
"non_date", or "blank".
}
\description{
This function evaluates each element of a character vector \code{x} to determine
if it represents a date. It is designed to recognize various numeric and
text-based date formats.
}
\keyword{internal}
