% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_preprocess.R
\name{core_prep_check_str_logical}
\alias{core_prep_check_str_logical}
\title{Check if a character vector contains logical formats.}
\usage{
core_prep_check_str_logical(x, allowed_strings = NULL)
}
\arguments{
\item{x}{A character vector to be checked.}

\item{allowed_strings}{A character vector of strings to be ignored during the
check. (These are called "allowed strings" as they are considered part of
a logical expression, e.g., "Response:")}
}
\value{
A character vector of the same length as \code{x} with values "logical",
"non_logical", or "blank".
}
\description{
This function evaluates each element of a character vector \code{x} to determine
if it represents a logical value (e.g., TRUE, FALSE, T, F, yes, no).
}
\keyword{internal}
