% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_preprocess.R
\name{core_prep_check_str_num}
\alias{core_prep_check_str_num}
\title{Check if a character vector contains numbers based on custom rules.}
\usage{
core_prep_check_str_num(
  x,
  allowed_strings = NULL,
  exclude_dates = TRUE,
  use_num_regex = c("most_relaxed", "relaxed", "strict")
)
}
\arguments{
\item{x}{A character vector to be checked.}

\item{allowed_strings}{A character vector of strings to be ignored during the
check. (These are called "allowed strings" as these are considered part of
number if supplied with at least one number)}

\item{exclude_dates}{A logical value. If \code{TRUE} (the default), strings that
resemble dates are classified as "non_num".}

\item{use_num_regex}{A character string indicating the regex pattern to use
for numbers should be one of: "most_relaxed", "relaxed", or "strict".
Defaults to "most_relaxed".}
}
\value{
A character vector of the same length as \code{x} with values "num",
"non_num", or "blank".
}
\description{
This function evaluates each element of a character vector \code{x} to determine
if it's a number ("num"), not a number ("non_num"), or effectively blank
("blank").
}
\keyword{internal}
