% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_preprocess.R
\name{core_prep_check_str_time}
\alias{core_prep_check_str_time}
\title{Check if a character vector contains time formats.}
\usage{
core_prep_check_str_time(
  x,
  allowed_strings = NULL,
  exclude_numbers = TRUE,
  use_date_regex = c("base", "base_with_qtrs", "base_with_qtrs_with_fy",
    "base_with_qtrs_with_fy_extra")
)
}
\arguments{
\item{x}{A character vector to be checked.}

\item{allowed_strings}{A character vector of strings to be ignored during the
check. (These are called "allowed strings" as they are considered part of
a time expression, e.g., "Starts at")}

\item{exclude_numbers}{Logical, default TRUE. If TRUE, then pure number strings
(e.g. "10.4", "123") are excluded from being classified as "time" even if they
match a numeric time pattern.}
}
\value{
A character vector of the same length as \code{x} with values "time",
"non_time", or "blank".
}
\description{
This function evaluates each element of a character vector \code{x} to determine
if it represents a time, optionally preceded by a date. Supports numeric times
(e.g., "10:40", "7pm"), multi-letter time units (e.g., "4hr", "10min", "80seconds"),
and combinations of at least two time units (e.g., "10h 20m", "30m 20s", "10h 20m 30s")
with single-letter, three-letter or full-word forms. If a value is a pure number (e.g., "10.4"),
it will be classified as "non_time" if \code{exclude_numbers = TRUE}.
}
\keyword{internal}
