% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes.R
\name{core_validate_cells}
\alias{core_validate_cells}
\title{Validate a cells object}
\usage{
core_validate_cells(dat)
}
\arguments{
\item{dat}{An R object to validate against the \code{cells} definition.}
}
\value{
A logical scalar. If \code{FALSE}, it includes a \code{msg} attribute detailing
all reasons for validation failure.
}
\description{
Validate a cells object
}
\details{
This function performs a comprehensive check to ensure an object
conforms to the \code{cells} structure. A valid \code{cells} object has the following
properties:
\itemize{
\item It is a data.frame that conforms to the \code{rc_df} format.
\item It has required character columns: \code{row}, \code{col}, \code{data_type}, \code{value}.
\item The \code{data_type} column must only contain values from core_cell_recognized_format().
\item The \code{value} column must always be a character string.
\item Optional \code{type} column: If present, must also only contain values from core_cell_recognized_format().
\item If \code{type} is present, \code{PoA} and \code{PoV} columns must also be present and valid.
\item Optional \code{sheet} column: If present, must be atomic, non-NA, and contain only one unique value.
}
}
\keyword{internal}
