% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_direction.R
\name{get_direction}
\alias{get_direction}
\title{get optimal direction (in terms of \code{unpivotr} directions)}
\usage{
get_direction(d_part)
}
\arguments{
\item{d_part}{parts of \code{data_attr_map_raw}}
}
\value{
a string value denoting direction
}
\description{
get optimal direction (in terms of \code{unpivotr} directions)
}
\details{
Used internally
}
\keyword{internal}
