% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_direction_df.R
\name{get_direction_df}
\alias{get_direction_df}
\title{Get Directional Orientation for Attributes corresponding to a given data
group}
\usage{
get_direction_df(dp, datt, allow_inside = FALSE)
}
\arguments{
\item{dp}{single data group boundary information}

\item{datt}{attribute data (including attribute group id)}
}
\value{
Directional Orientation for Attributes
}
\description{
Get Directional Orientation for Attributes corresponding to a given data
group
}
\details{
Used internally by get_data_block_information function
}
\keyword{internal}
