% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_group_id.R
\name{get_group_id}
\alias{get_group_id}
\title{Get Group ID for joined cells}
\usage{
get_group_id(dat)
}
\arguments{
\item{dat}{the (data or attribute) cells (in at least row-col format)}
}
\value{
Group ID attached information in a list
}
\description{
Get Group ID for joined cells
}
\details{
Used internally by get_data_block_information function
}
\keyword{internal}
