% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_unpivotr_direction_names.R
\name{get_unpivotr_direction_names}
\alias{get_unpivotr_direction_names}
\title{get unpivotr direction names}
\usage{
get_unpivotr_direction_names()
}
\value{
directions as used in unpivotr package with directional grouping
}
\description{
get unpivotr direction names
}
\details{
Kept for compatibility. Used internally by get_direction function.
}
\keyword{internal}
