% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_helpers_header_orientation_tag.R
\name{infer_get_HOT_metric}
\alias{infer_get_HOT_metric}
\title{Infer HOT (header_orientation_tag) metric}
\usage{
infer_get_HOT_metric(data_cells, header_cells, direction)
}
\arguments{
\item{data_cells}{Data frame of data block cells (with data_gid)}

\item{header_cells}{Data frame of header/attribute cells (with attr_gid)}

\item{direction}{Direction name (should match one of those from infer_get_valid_header_orientation_tags())}
}
\value{
A numeric value between 0 and 1: fraction of data_cells covered by the header in the given direction.
}
\description{
Computes a scaled fraction denoting the coverage (1 means full coverage) for the supplied direction.
}
\keyword{internal}
