% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pkg_availability_sim.R
\name{pkg_is_available}
\alias{pkg_is_available}
\title{Check Package Availability}
\usage{
pkg_is_available(pkg, check_working = FALSE)
}
\arguments{
\item{pkg}{Character string. Name of the package to check.}

\item{check_working}{Logical. If TRUE, also verifies functionality using a
custom-defined check (if available).}
}
\value{
Logical. TRUE if available (and working, if checked); FALSE
otherwise.
}
\description{
Determines if a package is installed and optionally functional for the host
package.
}
\seealso{
\code{\link[=pkg_not_available]{pkg_not_available()}}
}
\keyword{internal}
