% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{plot.cells}
\alias{plot.cells}
\title{Plot a Cells Object as a Grid}
\usage{
\method{plot}{cells}(
  x,
  ...,
  fill,
  fill_alpha = 1,
  ignore_validation = FALSE,
  no_fill = FALSE,
  txt_size = 3,
  txt_alpha = 1,
  no_txt = FALSE,
  max_txt_len = 10,
  no_plot = FALSE,
  boundaries = NULL,
  add_cell_address = FALSE,
  no_legend = TRUE,
  cell_connection_map = tibble::tibble(),
  connection_line_type = c("L-shaped", "Curved"),
  connection_line_alpha = 0.7,
  connection_line_add_jitter = TRUE,
  connection_line_overlay_text_color = TRUE,
  allow_rc_df = FALSE,
  rc_df_auto_txt_n_fill = TRUE,
  omit_NSE = FALSE,
  display_warning_as_plot = FALSE,
  scale_fill_values = NULL,
  auto_round_values = FALSE,
  declutter_based_on_value = FALSE,
  unify_on_row_col = FALSE
)
}
\arguments{
\item{x}{A \code{Cells} object to be plotted.}

\item{...}{Additional arguments passed. (currently unused)}

\item{fill}{Character. Name of the column in \code{x} to use for tile fill
colors. If missing, uses "type" if present, otherwise "data_type".}

\item{fill_alpha}{Numeric. Alpha transparency for the fill color of tiles.}

\item{ignore_validation}{Logical. If \code{TRUE}, skips validation of the
object before plotting. Default is \code{FALSE}.}

\item{no_fill}{Logical. If \code{TRUE}, disables fill coloring for tiles.
Default is \code{FALSE}.}

\item{txt_size}{Numeric. Text size for cell labels.}

\item{txt_alpha}{Numeric. Alpha transparency for cell labels.}

\item{no_txt}{Logical. If \code{TRUE}, disables text labels on the grid.
Default is \code{FALSE}.}

\item{max_txt_len}{Integer. Maximum length of text displayed in each cell
before truncation. Default is \code{10}.}

\item{no_plot}{Logical. If \code{TRUE}, returns the ggplot object without
displaying / plotting it. Default is \code{FALSE}.}

\item{boundaries}{Data frame containing boundary rectangles to be drawn on
the grid. Should have columns \code{c_min}, \code{c_max}, \code{r_min}, \code{r_max} for
coordinates, and optionally a grouping variable.}

\item{add_cell_address}{Logical. If \code{TRUE}, adds cell addresses as text
labels on the grid. Default is \code{FALSE}.}

\item{no_legend}{Logical. If \code{TRUE}, disables the legend in the plot.
Default is \code{TRUE}.}

\item{cell_connection_map}{Data frame containing information about cell
connections (required columns: \code{col_from}, \code{row_from}, \code{col_to}, \code{row_to};
optional column: \code{color}). Used to draw arrows between cells. This is
mainly for inspecting how cells are connected post composition or collate
stage.}

\item{connection_line_type}{Character. Type of connection lines to draw.
Options are "L-shaped" or "Curved". Default is "L-shaped".}

\item{connection_line_alpha}{Numeric. Alpha transparency for the connection
lines between cells. Default is \code{0.7}.}

\item{connection_line_add_jitter}{Logical. If \code{TRUE}, adds slight random
jitter to the start and end points of connection lines to reduce overlap.
Default is \code{TRUE}.}

\item{connection_line_overlay_text_color}{Logical. If \code{TRUE}, overlays
text color on top of connection lines based on the \code{color} column in
\code{cell_connection_map}. Default is \code{TRUE}.}

\item{allow_rc_df}{Logical. If \code{TRUE}, allows the input to be any rc_df.
This is designed for mainly testing and development purposes. Default is
\code{FALSE}.}

\item{omit_NSE}{Logical. If \code{TRUE}, treats the \code{fill} argument as a
character string rather than a symbol. This is mainly for internal use to
omit non-standard evaluation (NSE) in the fill argument. Default is
\code{FALSE}.}

\item{display_warning_as_plot}{Logical. If \code{TRUE}, displays a warning
plot instead of a text warning when the object has more than 200 rows or
columns. Default is \code{FALSE}. This is mainly for shiny modules or for
displaying warnings in a more user-friendly way.}

\item{scale_fill_values}{Named character vector. If provided, specifies the
fill colors for specific types or data types in the plot. This allows to
customize the color mapping for specific fill types, especially for plot of
cells-analysis.}
}
\value{
Invisibly returns the \code{ggplot} object for further customization
or display.
}
\description{
Visualizes a \code{Cells} object as a grid using \code{ggplot2}, where each
cell is represented by a tile. The fill color and text labels can be
customized.
}
\details{
\itemize{
\item Requires the \code{ggplot2} package.
\item Validates the \code{Cells} object unless \code{ignore_validation} is \code{TRUE}.
\item Warns and does not do anything if the object has more than 200 rows or columns.
\item If the specified \code{fill} column is not found in \code{x}, the function aborts with an error.
\item Text in cells longer than \code{max_txt_len} is truncated and appended with "...".
\item The function returns (invisibly) the \code{ggplot} object.
}
}
\keyword{internal}
