% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{print.cells}
\alias{print.cells}
\title{Print Summary for Cells Object}
\usage{
\method{print}{cells}(x, ignore_validation = FALSE, xl_view = TRUE, ...)
}
\arguments{
\item{x}{A Cells object.}

\item{ignore_validation}{Logical. If TRUE, skips validation of the object.
Default is FALSE.}

\item{xl_view}{Logical. If TRUE (default), displays a Excel - styled glimpse
of the cell data (if required packages are available).}

\item{...}{Further arguments passed to methods (currently unused).}
}
\description{
Prints a summary of a Cells object, including the number of cells and their
dimensions. Optionally glances at the data in a styled view and validates the
object before printing.
}
\keyword{internal}
