% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{print.cells_analysis}
\alias{print.cells_analysis}
\title{Print a Cells Analysis Object}
\usage{
\method{print}{cells_analysis}(x, ...)
}
\arguments{
\item{x}{A Output of cells/sheet analysis.}

\item{...}{Further arguments passed to methods (currently unused).}
}
\description{
Print a Cells Analysis Object
}
\keyword{internal}
