% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{print.cells_composition}
\alias{print.cells_composition}
\title{Print a Cells Composition Object}
\usage{
\method{print}{cells_composition}(x, ...)
}
\arguments{
\item{x}{A Cells Composition object.}

\item{...}{Further arguments passed to methods (currently unused).}
}
\value{
Invisibly returns the Cells Composition object.
}
\description{
Prints a summary of the Cells Composition object, including the number of
node/data/value-blocks.
}
\keyword{internal}
