% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class-read_cell_part.R
\name{read_cell_part-class}
\alias{read_cell_part-class}
\title{\code{read_cell_part} class}
\description{
The \code{read_cell_part} class is based on list,
created in order to store (possible) intermediate \code{\link[=read_cells]{read_cells}} output in detailed format.

\strong{Note:} This class is designed mainly for \emph{internal use}.
It is recommended to manipulate objects of this class only for cases where desired output is not coming.
}
\section{Properties of \code{read_cell_part}}{


Objects of class \code{read_cell_part} \strong{may} have following named nodes (the nodes may change based on stage):
\itemize{
\item \code{file_name} : file name which was given to process
\item \code{stage} : stage at which it was last processed.
\item \code{info} : a list containing further information on type and content of the file (provided the file got read)
\item \code{is_empty} : whether the file contains no 'known tabular information'
\item \code{cell_list} : list of \code{\link[=cell_df-class]{cell_df}}
(possibly after \verb{Value Attribute Classification} done if stage is higher than \code{make_cells})
\item \code{cell_analysis_list} : list of \code{\link[=cell_analysis-class]{cell_analysis}}
\item \code{final_composition} : final composition (a \code{tibble}) with only main columns
(or all column if \code{compose_main_cols_only = FALSE})
\item \code{final_composition_main} : only appear if \code{compose_main_cols_only = FALSE}.
This holds final composition (a \code{tibble}) with only main columns
}
}

\section{Applicable methods on \code{read_cell_part}}{

\itemize{
\item \code{print}: Prints identifier of the class and the stage at which it is last processed.
}
}

\keyword{internal}
