% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cells.R
\name{read_cells.read_cell_part}
\alias{read_cells.read_cell_part}
\alias{read_cells.character}
\alias{read_cells.default}
\alias{read_cells.NULL}
\title{Read Cells from file}
\usage{
\method{read_cells}{read_cell_part}(x, at_level = c("collate",
  "detect_and_read", "make_cells", "va_classify", "analyze", "compose"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE,
  from_level, silent = TRUE, ...)

\method{read_cells}{character}(x, at_level = c("collate",
  "detect_and_read", "make_cells", "va_classify", "analyze", "compose"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE, ...)

\method{read_cells}{default}(x, at_level = c("collate",
  "detect_and_read", "make_cells", "va_classify", "analyze", "compose"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE,
  from_level, ...)

\method{read_cells}{NULL}(x, ...)
}
\arguments{
\item{x}{either a valid file path or a \code{\link[=read_cell_part-class]{read_cell_part}}}

\item{at_level}{till which level to process.
Should be one of \code{detect_and_read}, \code{make_cells}, \code{va_classify}, \code{analyze}, \code{compose}. Or simply a number.}

\item{omit}{(Optional) the file-types to omit. A character vector.}

\item{simplify}{whether to simplify the output. (Default \code{TRUE}). If set to \code{FALSE} a \code{\link[=read_cell_part-class]{read_cell_part}}
will be returned.}

\item{compose_main_cols_only}{whether to compose main columns only. (Default \code{TRUE}).}

\item{from_level}{(Optional) override start level. (\code{read_cells} will process after \code{from_level})}

\item{silent}{if \code{TRUE} no message will be displayed.(Default \code{TRUE})}

\item{...}{further arguments}
}
\description{
Read Cells from file
}
\seealso{
\code{\link[=read_cells]{read_cells}}
}
\keyword{internal}
