% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options_public.R
\name{set_options}
\alias{set_options}
\title{Set Package Options}
\usage{
set_options(
  derive_nice_names_for_attributes = TRUE,
  excel_specific_optimization = FALSE,
  check_attr_entrapment_on_full_overlap_join = TRUE,
  finalize_through_stages = FALSE,
  finalize_till = "collate_column",
  collate_columns_on_whole = TRUE,
  auto_detach_shiny = TRUE,
  delink_corner_headers = FALSE,
  threshold_complete_connect_attempt_partial_gid_joins = 1000,
  focus_value_range = c(-Inf, Inf),
  focus_absolute_value_range = c(0, Inf),
  single_data_chunk_filter = FALSE,
  single_data_chunk_value_capture_threshold = 0.75
)
}
\arguments{
\item{derive_nice_names_for_attributes}{Logical. Should the package derive
user-friendly names for attributes? Default is \code{TRUE}.}

\item{excel_specific_optimization}{Logical. Should Excel-specific
optimization be enabled? Default is \code{FALSE}.}

\item{check_attr_entrapment_on_full_overlap_join}{Logical. Should the package
check for attribute entrapment during full overlap joins? Default is
\code{FALSE}.}

\item{finalize_through_stages}{Logical. Should the package finalize through
multiple stages? Default is \code{FALSE}.}

\item{finalize_till}{Character. Specifies the stage till which to finalize.
Options are "collate_column" (default), "compose", "analyze_cells", and
"value_attribute_classify".}

\item{collate_columns_on_whole}{Logical. Should columns be collated on whole?
Default is \code{TRUE}.}

\item{auto_detach_shiny}{Logical. Should the package automatically detach
Shiny when done? Default is \code{TRUE}.}

\item{delink_corner_headers}{Logical. Should corner headers be delinked from
data blocks? Default is \code{FALSE}.}

\item{threshold_complete_connect_attempt_partial_gid_joins}{Numeric. Threshold
limit for complete connect attempts on partial gid joins. Default is \code{1000}.}

\item{focus_value_range}{Numeric vector of length 2. Defines the min and max
value range to focus on. Default is \code{c(-Inf, Inf)}.}

\item{focus_absolute_value_range}{Numeric vector of length 2. Defines the min
and max absolute value range to focus on. Default is \code{c(0, Inf)}.}

\item{single_data_chunk_filter}{Logical. Should filtering for a single data
chunk be applied? Default is \code{FALSE}.}

\item{single_data_chunk_value_capture_threshold}{Numeric. Threshold proportion
for capturing values in a single data chunk. Default is \code{0.75}.}
}
\value{
No return value; (called for side effects - setting user parameters.)
}
\description{
This function sets user-specific options for the package, such as whether to
derive nice attribute names or enable Excel-specific optimization etc..
}
\details{
This function stores the options in the package cache under the
"user" namespace, which can affect package behavior such as naming
conventions and optimization for Excel based cells etc.
}
