% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{util_make_unique_minimal}
\alias{util_make_unique_minimal}
\title{Make Vector Elements Unique with Minimal Suffixes}
\usage{
util_make_unique_minimal(x)
}
\arguments{
\item{x}{A character or numeric vector.}
}
\value{
A character vector the same length as \code{x}, with duplicates made
unique by the minimal addition of integer suffixes.
}
\description{
Adds numeric suffixes to duplicated elements in a character or numeric
vector, using the minimal number of changes required to ensure all elements
are unique. The first occurrence of each unique value is left unchanged;
subsequent duplicates receive a "_2", "_3", etc. suffix.
}
\examples{
\dontrun{
util_make_unique_minimal(c("a", "b", "a", "a", "c", "b"))
# [1] "a" "b" "a_2" "a_3" "c" "b_2"

util_make_unique_minimal(c(1, 2, 2, 3, 1))
# [1] "1" "2" "2_2" "3" "1_2"
}

}
\keyword{internal}
