% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{util_most_frequent}
\alias{util_most_frequent}
\title{Infer the most frequent character in a vector}
\usage{
util_most_frequent(x)
}
\arguments{
\item{x}{A character vector.}
}
\value{
The most frequent character in the vector, or \code{NA_character_} if the
vector is empty.
}
\description{
This function takes a character vector and returns the most frequently occurring
character. If there are multiple characters with the same maximum frequency,
it returns the first one in alphabetical order. If the input vector is empty,
it returns \code{NA_character_}.
}
\keyword{internal}
