% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_utils.R
\name{util_vector_operation_optim}
\alias{util_vector_operation_optim}
\title{Optimally apply a function to a vector by handling duplicates}
\usage{
util_vector_operation_optim(
  x,
  f,
  ratio_threshold = 0.7,
  length_threshold = 10000,
  paired_arguments = NULL,
  ...
)
}
\arguments{
\item{x}{Input vector.}

\item{f}{A vectorized function to apply to \code{x}. Must be "value-based" (see
Details).}

\item{ratio_threshold}{If (\code{n_distinct / n_total}) is below this, the
optimization is triggered. Defaults to 0.7.}

\item{length_threshold}{If \code{length(x)} is below this, \code{f} is applied
directly. Defaults to 10000.}

\item{paired_arguments}{Optional character vector of additional argument
names that, if supplied, should be subset in parallel with unique values of
\code{x}.}

\item{...}{Additional arguments passed to \code{f}.}
}
\value{
A vector with the result of \code{f(x, ...)}.
}
\description{
Efficiently applies a (potentially expensive) function to a vector by
avoiding redundant computations for duplicate values. If the input vector is
large and contains many repeated values, the function is applied only to the
unique values, and results are mapped back to the original ordering. This
optimization is triggered only when both the input length and the proportion
of unique values cross user-specified thresholds.
}
\details{
This function is only safe for "value-based" functions: those for which
\code{f(x)[i]} depends only on the value of \code{x[i]}, not on its position in \code{x} or
any global state. The output of \code{f(x)} must be the same length as \code{x}.

If the function also requires other arguments that should be subset in sync
with \code{x}, their names may be supplied in \code{paired_arguments}.
}
\keyword{internal}
