% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/visual_functions.R
\name{visual_value_attribute_classify}
\alias{visual_value_attribute_classify}
\alias{visual_orientation_modification}
\alias{visual_traceback}
\title{Interactive operations and Visualizations}
\usage{
visual_value_attribute_classify(d)

visual_orientation_modification(x)

visual_traceback(x)
}
\arguments{
\item{d}{a cells object.}

\item{x}{a cells-analysis object.}
}
\value{
\code{visual_value_attribute_classify} returns a cell_df after the
interactive operations.

\code{visual_orientation_modification} returns a cell_analysis after the
interactive operations.

\code{visual_traceback} returns a composed object after interactive operations.
}
\description{
These functions require \code{\link[shiny:shiny-package]{shiny}}
and \code{miniUI}.
}
\details{
Each of these functions are designed for interactive sessions only.
Arguments to these functions are optional as it tries to pick those from
global environment. If there are multiple variables defined in global
environment which are of same desired class then you will get option to
choose one of them. These are also available as
\href{https://rstudio.github.io/rstudioaddins/}{RStudio Addin} (if you are
using RStudio IDE you'll get those addin in the RStudio addin panel).

Following are the names used in RStudio Addin
\itemize{
\item \emph{Orientation Modification}: \code{visual_orientation_modification}
\item \emph{Traceback Composition}: \code{visual_traceback}
\item \emph{Classify Value/Attribute}: \code{visual_value_attribute_classify}

For each of the modules it tries to load and unload/detach \code{shiny} before
and after the execution of the module (or rather it will honour the state
in which theshiny was before and after the function. Means if shiny was
atatched then it will remain attahced otherwise will be detached.
}
}
\seealso{
\code{\link{value_attribute_classify}}
}
\keyword{internal}
