% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analyze_cells.R
\name{analyze_cells}
\alias{analyze_cells}
\title{Analyze Cells}
\usage{
analyze_cells(d, silent = TRUE)
}
\arguments{
\item{d}{A cell_df after \code{Value Attribute Classification} done}

\item{silent}{logical scalar indicating whether to raise a warning if heuristic detection fails. (Default TRUE).}
}
\value{
Detailed analysis of the cell data structure
}
\description{
Analyze Cells
}
\details{
it returns detailed analysis of the data structure including data block detection, attribute orientation detection etc.
The argument \code{silent} is set to \code{TRUE} by default, as the warning will be given whenever the cell-analysis is printed.
}
\examples{
d <- structure(c(
  "block 1", "", "C", "D", "", "block 2", "", "C",
  "D", "", "A", "1", "2", "", "", "A", "10", "20", "", "B", "3",
  "4", "", "", "B", "30", "40"
), .Dim = c(9L, 3L))
d <- as.data.frame(d)
cd <- as_cell_df(d) \%>\% numeric_values_classifier()

# see it
cd \%>\% plot(adaptive_txt_size = FALSE)
ca <- analyze_cells(cd)

# look at the plot for detected directions
plot(ca)
}
