% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_cell_df.R
\name{as_cell_df}
\alias{as_cell_df}
\title{Transform data into Cell-DF Structure}
\usage{
as_cell_df(d, take_row_names = FALSE, take_col_names = FALSE)
}
\arguments{
\item{d}{the data (either a matrix with column name or a data.frame)}

\item{take_row_names}{consider row names as separate cells
(applicable only for data with no (row, col) information). Default is \code{FALSE}.}

\item{take_col_names}{consider column names as separate cells
(applicable only for data with no (row, col) information). Default is \code{FALSE}.}
}
\value{
A object of class \code{\link[=cell_df-class]{cell_df}}.
}
\description{
Transform data into Cell-DF Structure
}
\examples{

as_cell_df(iris)

# consider column name as cell
as_cell_df(iris, take_col_names = TRUE)

# if the data is already in similar format it will not further transform
unpivotr::as_cells(iris) \%>\% as_cell_df()
}
