% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/collate_columns.R
\name{collate_columns}
\alias{collate_columns}
\title{Collate Columns Based on Content}
\usage{
collate_columns(composed_data, combine_threshold = 0.9,
  rest_cols = Inf, retain_other_cols = FALSE,
  retain_cell_address = FALSE)
}
\arguments{
\item{composed_data}{output of \code{\link[=compose_cells]{compose_cells}} (preferably not processed)}

\item{combine_threshold}{a numerical threshold (between 0-1) for content-based collation of columns. (Default 0.9)}

\item{rest_cols}{number of rest columns (beyond \code{combine_threshold} joins these many number of columns to keep)}

\item{retain_other_cols}{whether to keep other intermediate (and possibly not so important) columns. (Default \code{FALSE})}

\item{retain_cell_address}{whether to keep columns like (\code{row}, \code{col}, \code{data_block}).
This may be required for \code{\link[=cell_composition_traceback]{traceback}} (Default \code{FALSE})}
}
\value{
A column collated data.frame
}
\description{
Collate Columns Based on Content
}
\examples{

d <- system.file("extdata", "marks_cells.rds", package = "tidycells", mustWork = TRUE) \%>\%
  readRDS()
d <- numeric_values_classifier(d)
da <- analyze_cells(d)

dc <- compose_cells(da, print_attribute_overview = TRUE)

collate_columns(dc)
}
