% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_classes_generics.R
\name{plot.cells}
\alias{plot.cells}
\title{Plot a Cells Object as a Grid}
\usage{
\method{plot}{cells}(
  x,
  ...,
  fill,
  ignore_validation = FALSE,
  no_fill = FALSE,
  txt_size = 3,
  txt_alpha = 1,
  no_txt = FALSE,
  max_txt_len = 10,
  no_plot = FALSE,
  boundaries = NULL,
  add_cell_address = FALSE,
  allow_rc_df = FALSE
)
}
\arguments{
\item{x}{A \code{Cells} object to be plotted.}

\item{...}{Additional arguments passed. (currently unused)}

\item{fill}{Character. Name of the column in \code{x} to use for tile fill
colors. If missing, uses "type" if present, otherwise "data_type".}

\item{ignore_validation}{Logical. If \code{TRUE}, skips validation of the
object before plotting. Default is \code{FALSE}.}

\item{no_fill}{Logical. If \code{TRUE}, disables fill coloring for tiles.
Default is \code{FALSE}.}

\item{txt_size}{Numeric. Text size for cell labels.}

\item{txt_alpha}{Numeric. Alpha transparency for cell labels.}

\item{no_txt}{Logical. If \code{TRUE}, disables text labels on the grid.
Default is \code{FALSE}.}

\item{max_txt_len}{Integer. Maximum length of text displayed in each cell
before truncation. Default is \code{10}.}

\item{no_plot}{Logical. If \code{TRUE}, returns the ggplot object without
displaying / plotting it. Default is \code{FALSE}.}

\item{boundaries}{Data frame containing boundary rectangles to be drawn on
the grid. Should have columns \code{c_min}, \code{c_max}, \code{r_min}, \code{r_max} for
coordinates, and optionally a grouping variable.}

\item{add_cell_address}{Logical. If \code{TRUE}, adds cell addresses as text
labels on the grid. Default is \code{FALSE}.}

\item{allow_rc_df}{Logical. If \code{TRUE}, allows the input to be any rc_df.
This is designed for mainly testing and development purposes. Default is
\code{FALSE}.}
}
\value{
Invisibly returns the \code{ggplot} object for further customization
or display.
}
\description{
Visualizes a \code{Cells} object as a grid using \code{ggplot2}, where each
cell is represented by a tile. The fill color and text labels can be
customized.
}
\details{
\itemize{
\item Requires the \code{ggplot2} package.
\item Validates the \code{Cells} object unless \code{ignore_validation} is \code{TRUE}.
\item Warns and does not do anything if the object has more than 200 rows or columns.
\item If the specified \code{fill} column is not found in \code{x}, the function aborts with an error.
\item Text in cells longer than \code{max_txt_len} is truncated and appended with "...".
\item The function returns (invisibly) the \code{ggplot} object.
}
}
\keyword{internal}
