% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_cells.R
\name{read_cells}
\alias{read_cells}
\alias{read_cells.read_cell_part}
\alias{read_cells.character}
\alias{read_cells.default}
\alias{read_cells.NULL}
\title{Read Cells from file}
\usage{
read_cells(x, at_level = c("compose", "detect_and_read", "make_cells",
  "va_classify", "analyze"), omit = NULL, simplify = TRUE,
  compose_main_cols_only = TRUE, from_level, silent = TRUE, ...)

\method{read_cells}{read_cell_part}(x, at_level = c("compose",
  "detect_and_read", "make_cells", "va_classify", "analyze"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE,
  from_level, silent = TRUE, ...)

\method{read_cells}{character}(x, at_level = c("compose",
  "detect_and_read", "make_cells", "va_classify", "analyze"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE, ...)

\method{read_cells}{default}(x, at_level = c("compose",
  "detect_and_read", "make_cells", "va_classify", "analyze"),
  omit = NULL, simplify = TRUE, compose_main_cols_only = TRUE,
  from_level, ...)

\method{read_cells}{NULL}(x, ...)
}
\arguments{
\item{x}{either a valid file path or a \code{\link[=read_cell_part-class]{read_cell_part}}}

\item{at_level}{till which level to process.
Should be one of \code{detect_and_read}, \code{make_cells}, \code{va_classify}, \code{analyze}, \code{compose}. Or simply a number.}

\item{omit}{(Optional) the file-types to omit. A character vector.}

\item{simplify}{whether to simplify the output. (Default \code{TRUE}). If set to \code{FALSE} a \code{\link[=read_cell_part-class]{read_cell_part}}
will be returned.}

\item{compose_main_cols_only}{whether to compose main columns only. (Default \code{TRUE}).}

\item{from_level}{(Optional) override start level. (\code{read_cells} will process after \code{from_level})}

\item{silent}{if \code{TRUE} no message will be displayed.(Default \code{TRUE})}

\item{...}{further arguments}
}
\value{
If \code{simplify=TRUE} then different kind of object is returned in different levels (depends on \code{at_level}).
If \code{at_level="compose"} then only final tibble is returned otherwise if the output is not \code{NULL} an attribute will be present
named \code{"read_cells_stage"}.

If \code{simplify=FALSE} then it will return a \code{\link[=read_cell_part-class]{read_cell_part}} which you can process manually
and continue again with \code{read_cells} (perhaps then \code{from_level} may be useful).
}
\description{
This function is designed to read cell level information
(and the finally \link[=analyze_cells]{analyze} and \link[=compose_cells]{compose}) from many file types like xls, pdf, doc etc.
This is a wrapper function to functions from multiple packages. The support for specific file is dependent on
the installed packages. To see the list of supported files and potentially required packages (if any) just
run \code{read_cells()} in the console. This function supports the file format based on content and not based on just the file
extension. That means if a file is saved ad pdf and then extension is removed (or extension modified to say \code{.xlsx})
then also the \code{read_cells} will detect it as pdf and read its content.
}
\examples{
# see supported files
read_cells()
}
