% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/core_options_public.R
\name{set_options}
\alias{set_options}
\title{Set Package Options}
\usage{
set_options(
  derive_nice_names_for_attributes = FALSE,
  excel_specific_optimization = FALSE
)
}
\arguments{
\item{derive_nice_names_for_attributes}{Logical. Should the package derive
user-friendly names for attributes? Default is \code{FALSE}.}

\item{excel_specific_optimization}{Logical. Should Excel-specific
optimization be enabled? Default is \code{FALSE}.}
}
\value{
No return value; called for its side effects.
}
\description{
This function sets user-specific options for the package, such as whether to
derive nice attribute names or enable Excel-specific optimization etc..
}
\details{
This function stores the options in the package cache under the
"user" namespace, which can affect package behavior such as naming
conventions and optimization for Excel based cells etc.
}
